Moved pybb profile model into models submodule
